impr: Breaking down the ryot_compass:drawing mod
